app: Support the creation of several windows
Fairly minimal things, but needed. There are still some decisions to
make, ie. should windows share the same GL contexts or not (or if it
should be selectable.

I think I'll make the windows share the same context so we don't have to
call MakeCurrent() per window and to be able to share textures/fbos
between windows (for instance a control window will be able to have a
preview of the render)

Signed-off-by: Damien Lespiau <damien.lespiau@gmail.com>